pyproject.toml
[tool.poetry]
name = "munsellcolorreproduction"
version = "0.1.0"
description = ""
authors = ["Chiang <chiang7h7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
colour-science = "^0.4.4"
matplotlib = "^3.8.2"
pymupdf = "^1.23.21"
easyocr = "^1.7.1"
opencv-python = "^4.9.0.80"
pandas = "^2.2.1"
scipy = "^1.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
